+++ experiments/surnameClassifier.json
@@ -1,15 +1,14 @@
 {
-  "data_file": "HOME/work/experiments/nlp/data/books/frankenstein_with_splits.csv",
-  "model_state_file": "modelCBOW.pth",
-  "save_dir": "HOME/work/experiments/nlp/models/cbow",
-  "logs": "HOME/work/experiments/nlp/models/cbow/logs/log1",
-  "glove_filepath": "HOME/work/experiments/nlp/data/glove/glove.6B.100d.txt",
-  "embedding_size": 100,
-  "dropout_p": 0.3,
+  "data_file": "HOME/work/experiments/nlp/data/surnames/surnames_with_splits.csv",
+  "model_state_file": "modelCNN.pth",
+  "save_dir": "HOME/work/experiments/nlp/models/surnames/",
+  "logs": "HOME/work/experiments/nlp/models//surnamesCNN/logs/log1",
+  "hidden_dim": 10,
   "seed": 1337,
   "lr": 0.001,
   "batch_size": 128,
-  "num_epochs": 1,
+  "num_epochs": 10,
+  "num_channels": 10,
   "early_stopping_criteria": 5,
   "cuda": true,
   "alpha": 0.01,
@@ -17,29 +16,20 @@
   "mode": "min",
   "factor": 0.5,
   "patience": 1,
-  "vectorizer": {
-    "_name": "CBOWVectorizer1"
-  },
-  "embeddings": {
-    "_name": "Embedding"
-  },
-  "embedding_hyper_params": {
-    "_name": "EmbeddingsHyperParams"
-  },
-  "embeddings2model_hyper_params": {
-    "_name": "EmbeddingtoModelHyperParams1"
+    "vectorizer": {
+    "_name": "SurnamesVectorizerCNN"
   },
   "dataset_hyper_params": {
     "_name": "DatasetHyperParams"
   },
   "dataset_splits": {
-    "_name": "CBOWDataset"
+    "_name": "SurnamesCNN"
   },
   "model": {
-    "_name": "CBOWClassifier1"
+    "_name": "SurnameClassifierCNN"
   },
   "model_hyper_params": {
-    "_name": "CBOWHyperParams"
+    "_name": "SurnameCNNHyperParams"
   },
   "model_params": {
     "_name": "ModelParams"
@@ -72,7 +62,7 @@
     ]
   },
   "predictor": {
-    "_name": "CBOWPredictor"
+    "_name": "SurnameCNNPredictor"
   },
   "predictor_hyper_params": {
     "_name": "PredictorHyperParams"
