+++ transfer_nlp/__init__.py
@@ -1,2 +1,1 @@
-# from transfer_nlp.models import cbow, cnn, rnn, perceptrons, generation, nmt
-from transfer_nlp.plugins import metrics, losses, generators, schedulers, optimizers, regularizers, helpers, trainers
+from transfer_nlp.plugins import metrics, regularizers, helpers, trainers
