--- transfer_nlp/experiments/mlp2.json
@@ -1,66 +1,0 @@
-{
-  "dataset": "HOME/work/experiments/nlp/data/surnames/surnames_with_splits.csv",
-  "vectorizer_file": "vectorizerMLP.json",
-  "model_state_file": "modelMLP.pth",
-  "save_dir": "HOME/work/experiments/nlp/models/surnames/",
-  "logs": "HOME/work/experiments/nlp/models/surnames/logs/log1",
-  "glove_filepath": "HOME/work/experiments/nlp/data/glove/glove.6B.100d.txt",
-  "use_glove": false,
-  "hidden_dim": 100,
-  "seed": 1337,
-  "lr": 0.001,
-  "batch_size": 128,
-  "num_epochs": 5,
-  "early_stopping_criteria": 5,
-  "cuda": true,
-  "catch_keyboard_interrupt": true,
-  "reload_from_files": false,
-  "expand_filepaths_to_save_dir": true,
-  "dataset_cls": "SurnamesDataset2",
-  "model": {
-    "modelName": "MultiLayerPerceptron2",
-    "modelParams": [
-      "input_dim",
-      "hidden_dim",
-      "output_dim"
-    ],
-    "modelInputs": [
-      "x_in"
-    ]
-  },
-  "loss": {
-    "lossName": "CrossEntropyLoss",
-    "lossParams": []
-  },
-  "Optimizer": {
-    "optimizerName": "Adam",
-    "optimizerParams": [
-      "params",
-      "lr"
-    ]
-  },
-  "Regularizer": {
-    "regularizerName": "L1",
-    "regularizerParams": [
-      "alpha"
-    ]
-  },
-  "scheduler": {
-    "schedulerName": "ReduceLROnPlateau",
-    "schedulerParams": [
-      "optimizer",
-      "mode",
-      "factor",
-      "patience"
-    ]
-  },
-  "alpha": 0.01,
-  "gradient_clipping": {"value": 0.25},
-  "mode": "min",
-  "factor": 0.5,
-  "patience": 1,
-  "batch_generator": "generate_batches",
-  "metrics": [
-    "compute_accuracy"
-  ]
-}
