--- transfer_nlp/experiments/newsClassifier.json
@@ -1,77 +1,0 @@
-{
-  "dataset": "/Users/petermartigny/work/experiments/nlp/data/ag_news/news_with_splits.csv",
-  "vectorizer_file": "vectorizer.json",
-  "model_state_file": "model.pth",
-  "save_dir": "/Users/petermartigny/work/experiments/nlp/models/ag_news",
-  "logs": "/Users/petermartigny/work/experiments/nlp/models/ag_news/logs/log1",
-  "glove_filepath": "/Users/petermartigny/work/experiments/nlp/data/glove/glove.6B.100d.txt",
-  "use_glove": false,
-    "load_model": true,
-  "embedding_size": 100,
-  "hidden_dim": 10,
-  "num_channels": 10,
-  "seed": 1337,
-  "lr": 0.001,
-  "dropout_p": 0.1,
-  "batch_size": 128,
-  "num_epochs": 10,
-  "early_stopping_criteria": 5,
-  "cuda": false,
-  "catch_keyboard_interrupt": true,
-  "reload_from_files": false,
-  "expand_filepaths_to_save_dir": true,
-  "dataset_cls": "NewsDataset",
-  "model": {
-    "modelName": "NewsClassifier",
-    "modelParams": [
-      "embedding_size",
-      "num_embeddings",
-      "num_channels",
-      "hidden_dim",
-      "num_classes",
-      "dropout_p",
-      "pretrained_embeddings",
-      "padding_idx"
-    ],
-    "modelInputs": [
-      "x_in"
-    ]
-  },
-  "loss": {
-    "lossName": "CrossEntropyLoss",
-    "lossParams": [
-      "weight"
-    ]
-  },
-  "Optimizer": {
-    "optimizerName": "Adam",
-    "optimizerParams": [
-      "params",
-      "lr"
-    ]
-  },
-  "Regularizer": {
-    "regularizerName": "L1",
-    "regularizerParams": [
-      "alpha"
-    ]
-  },
-  "scheduler": {
-    "schedulerName": "ReduceLROnPlateau",
-    "schedulerParams": [
-      "optimizer",
-      "mode",
-      "factor",
-      "patience"
-    ]
-  },
-  "alpha": 0.01,
-  "gradient_clipping": {"value": 0.25},
-  "mode": "min",
-  "factor": 0.5,
-  "patience": 1,
-  "batch_generator": "generate_batches",
-  "metrics": [
-    "compute_accuracy"
-  ]
-}
