--- transfer_nlp/experiments/surnamesGeneration.json
@@ -1,77 +1,0 @@
-{
-  "dataset": "/Users/petermartigny/work/experiments/nlp/data/surnames/surnames_with_splits.csv",
-  "vectorizer_file": "vectorizerGeneration.json",
-  "model_state_file": "modelGeneration.pth",
-  "save_dir": "/Users/petermartigny/work/experiments/nlp/models/model1_surname_generation",
-  "logs": "/Users/petermartigny/work/experiments/nlp/models/model1_surname_generation/logs/log1",
-  "glove_filepath": "/Users/petermartigny/work/experiments/nlp/data/glove/glove.6B.100d.txt",
-  "use_glove": false,
-  "embedding_size": 100,
-  "char_embedding_size": 50,
-  "rnn_hidden_size": 100,
-  "hidden_dim": 10,
-  "num_channels": 10,
-  "seed": 1337,
-  "lr": 0.001,
-  "dropout_p": 0.5,
-  "batch_size": 128,
-  "num_epochs": 10,
-  "early_stopping_criteria": 5,
-  "cuda": true,
-  "catch_keyboard_interrupt": true,
-  "reload_from_files": false,
-  "expand_filepaths_to_save_dir": true,
-  "dataset_cls": "SurnameDatasetGeneration",
-  "model": {
-    "modelName": "SurnameConditionedGenerationModel",
-    "modelParams": [
-      "char_embedding_size",
-      "char_vocab_size",
-      "num_nationalities",
-      "rnn_hidden_size",
-      "padding_idx",
-      "dropout_p",
-      "conditioned"
-    ],
-    "modelInputs": [
-      "x_in",
-      "nationality_index"
-    ]
-  },
-  "loss": {
-    "lossName": "SequenceLoss",
-    "lossParams": []
-  },
-  "Optimizer": {
-    "optimizerName": "Adam",
-    "optimizerParams": [
-      "params",
-      "lr"
-    ]
-  },
-  "Regularizer": {
-    "regularizerName": "L1",
-    "regularizerParams": [
-      "alpha"
-    ]
-  },
-  "scheduler": {
-    "schedulerName": "ReduceLROnPlateau",
-    "schedulerParams": [
-      "optimizer",
-      "mode",
-      "factor",
-      "patience"
-    ]
-  },
-  "alpha": 0.01,
-  "gradient_clipping": {"value": 0.25},
-  "conditioned": true,
-  "mode": "min",
-  "factor": 0.5,
-  "patience": 1,
-  "batch_generator": "generate_batches",
-  "metrics": [
-    "compute_accuracy_sequence"
-  ]
-}
