--- transfer_nlp/experiments/surnamesRNN.json
@@ -1,79 +1,0 @@
-{
-  "dataset": "/Users/petermartigny/work/experiments/nlp/data/surnames/surnames_with_splits.csv",
-  "vectorizer_file": "vectorizerRNN.json",
-  "model_state_file": "modelRNN.pth",
-  "save_dir": "/Users/petermartigny/work/experiments/nlp/models/surnames",
-  "logs": "/Users/petermartigny/work/experiments/nlp/models/surnames/logs/log1",
-  "glove_filepath": "/Users/petermartigny/work/experiments/nlp/data/glove/glove.6B.100d.txt",
-  "use_glove": false,
-  "load_model": true,
-  "embedding_size": 100,
-  "char_embedding_size": 100,
-  "rnn_hidden_size": 64,
-  "hidden_dim": 10,
-  "num_channels": 10,
-  "seed": 1337,
-  "lr": 0.001,
-  "dropout_p": 0.1,
-  "batch_size": 64,
-  "num_epochs": 5,
-  "early_stopping_criteria": 5,
-  "cuda": true,
-  "catch_keyboard_interrupt": true,
-  "reload_from_files": false,
-  "expand_filepaths_to_save_dir": true,
-  "dataset_cls": "SurnameDatasetRNN",
-  "model": {
-    "modelName": "SurnameClassifierRNN",
-    "modelParams": [
-      "embedding_size",
-      "num_embeddings",
-      "num_classes",
-      "rnn_hidden_size",
-      "padding_idx"
-    ],
-    "modelInputs": [
-      "x_in",
-      "x_lengths"
-    ]
-  },
-  "loss": {
-    "lossName": "CrossEntropyLoss",
-    "lossParams": [
-      "weight"
-    ]
-  },
-  "Optimizer": {
-    "optimizerName": "Adam",
-    "optimizerParams": [
-      "params",
-      "lr"
-    ]
-  },
-  "Regularizer": {
-    "regularizerName": "L1",
-    "regularizerParams": [
-      "alpha"
-    ]
-  },
-  "scheduler": {
-    "schedulerName": "ReduceLROnPlateau",
-    "schedulerParams": [
-      "optimizer",
-      "mode",
-      "factor",
-      "patience"
-    ]
-  },
-  "alpha": 0.01,
-  "gradient_clipping": {
-    "value": 0.25
-  },
-  "mode": "min",
-  "factor": 0.5,
-  "patience": 1,
-  "batch_generator": "generate_batches",
-  "metrics": [
-    "compute_accuracy"
-  ]
-}
