+++ transfer_nlp/models/nmt.py
@@ -10,12 +10,12 @@ from torch.nn.utils.rnn import pack_padded_sequence, pad_packed_sequence
 from transfer_nlp.common.utils import describe
 from transfer_nlp.loaders.vectorizers import NMTVectorizer
 from transfer_nlp.loaders.vocabulary import SequenceVocabulary
-from transfer_nlp.plugins.registry import register_model
+from transfer_nlp.plugins.config import register_plugin
 
 chencherry = bleu_score.SmoothingFunction()
 
 
-@register_model
+@register_plugin
 class NMTEncoder(nn.Module):
 
     def __init__(self, num_embeddings: int, embedding_size: int, rnn_hidden_size: int):
@@ -76,7 +76,7 @@ def terse_attention(encoder_state_vectors: torch.Tensor, query_vector: torch.Ten
     return context_vectors, vector_probabilities
 
 
-@register_model
+@register_plugin
 class NMTDecoder(nn.Module):
 
     def __init__(self, num_embeddings: int, embedding_size: int, rnn_hidden_size: int, bos_index: int):
@@ -175,7 +175,7 @@ class NMTDecoder(nn.Module):
         return output_vectors
 
 
-@register_model
+@register_plugin
 class NMTModel(nn.Module):
 
     def __init__(self, source_vocab_size: int, source_embedding_size: int,
@@ -222,7 +222,7 @@ def sentence_from_indices(indices: List[int], vocab: SequenceVocabulary, strict:
         return out
 
 
-@register_model
+@register_plugin
 class NMTSampler:
 
     def __init__(self, vectorizer: NMTVectorizer, model: NMTModel):
