--- transfer_nlp/plugins/activation_functions.py
@@ -1,5 +1,0 @@
-"""
-If you decide to implement your own activation functions, you should implement an nn.functional class.
-There are already the most used pytorch activation functions included in the registry so you should be good to go, but you can still
-implement yours and add the decorator @register_activation which is in transfer_nlp.plugins.registry.py
-"""
